Continued work on mobs
:skull:
