--- 2DGame/Assets/Scripts/Mobs/SlimeEnemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a7f4401447dc5984884a38877f18351c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
