--- 2DGame/Assets/Scripts/Mobs/StoneGolemBoss.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cfe3caa1c94352443bcd5a9a9a772cbc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
