+++ 2DGame/Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.rsp
@@ -392,8 +392,7 @@
 "Assets/Scripts/Mobs/Interfaces/IProjectile.cs"
 "Assets/Scripts/Mobs/SkeletonEnemy.cs"
 "Assets/Scripts/Mobs/SkeletonProjectile.cs"
-"Assets/Scripts/Mobs/SlimeEnemy.cs"
-"Assets/Scripts/Mobs/StoneGolemBoss.cs"
+"Assets/Scripts/Mobs/SlimeBoss.cs"
 "Assets/Scripts/Mobs/ZombieEnemy.cs"
 "Assets/Scripts/perlinNoiseTexture.cs"
 "Assets/Scripts/TerrainGeneration.cs"
