--- 2DGame/Logs/shadercompiler-UnityShaderCompiler.exe-5.log
@@ -1,7 +1,0 @@
-Base path: 'C:/Program Files/Unity/Hub/Editor/2023.1.9f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2023.1.9f1/Editor/Data/PlaybackEngines'
-Cmd: initializeCompiler
-
-Cmd: compileSnippet
-  insize=1098 file=Assets/DefaultResourcesExtra/Particles/Standard Unlit pass=SceneSelectionPass ppOnly=0 stripLineD=0 buildPlatform=19 rsLen=0 pKW=UNITY_ENABLE_REFLECTION_BUFFERS UNITY_USE_DITHER_MASK_FOR_ALPHABLENDED_SHADOWS UNITY_PBS_USE_BRDF1 UNITY_SPECCUBE_BOX_PROJECTION UNITY_SPECCUBE_BLENDING UNITY_ENABLE_DETAIL_NORMALMAP SHADER_API_DESKTOP UNITY_LIGHT_PROBE_PROXY_VOLUME UNITY_LIGHTMAP_FULL_HDR uKW= dKW=INSTANCING_ON PROCEDURAL_INSTANCING_ON _ALPHATEST_ON _REQUIRE_UV2 UNITY_NO_DXT5nm UNITY_FRAMEBUFFER_FETCH_AVAILABLE UNITY_METAL_SHADOWS_USE_POINT_FILTERING UNITY_NO_SCREENSPACE_SHADOWS UNITY_PBS_USE_BRDF2 UNITY_PBS_USE_BRDF3 UNITY_HARDWARE_TIER1 UNITY_HARDWARE_TIER2 UNITY_HARDWARE_TIER3 UNITY_COLORSPACE_GAMMA UNITY_LIGHTMAP_DLDR_ENCODING UNITY_LIGHTMAP_RGBM_ENCODING UNITY_VIRTUAL_TEXTURING UNITY_PRETRANSFORM_TO_DISPLAY_ORIENTATION UNITY_ASTC_NORMALMAP_ENCODING SHADER_API_GLES30 UNITY_UNIFIED_SHADER_PRECISION_MODEL flags=0 lang=0 type=Fragment platform=d3d11 reqs=33 mask=6 start=100 ok=1 outsize=314
-
-Cmd: shutdown
